Update to version 0.2.0 with -dryrun and -force

+++ sasbuild.ps1
@@ -1,14 +1,28 @@
+# v0.2.0, by Michael Stepner, 2017-07-21
 param(
-	[Parameter(Mandatory=$true)][string]$build
+	[Parameter(Mandatory=$true)][string]$build,
+	[switch]$force,
+	[switch]$dryrun
 )
 
+# Config
+$sas_exe = "C:\Program Files\SASHome\SASFoundation\9.3\sas.exe"
+$sas_cfg = "C:\Program Files\SASHome\SASFoundation\9.3\sasv9.cfg"
+
+# Code
 function runsas ($file) {
 	Write-Host ((Get-Date).ToString() + " Running: " + $file + ".sas")
-	& "C:\Program Files\SASHome\SASFoundation\9.3\sas.exe" -CONFIG "C:\Program Files\SASHome\SASFoundation\9.3\sasv9.cfg" -SYSIN ($file + ".sas") -LOG ($file + ".log") -PRINT ($file + ".lst") -NOSPLASH | Out-Null
-	if ($LastExitCode -gt 1) {
-		Move-Item -Path ($file + ".log") -Destination ($file + ".error.log") -Force
-		Write-Host ((Get-Date).ToString() + " Error encountered while running " + $file + ".sas")
-		exit $LastExitCode
+	if (!$dryrun) {
+		& $sas_exe -CONFIG $sas_cfg -SYSIN ($file + ".sas") -LOG ($file + ".log") -PRINT ($file + ".lst") -NOSPLASH | Out-Null
+		if ($LastExitCode -gt 1) {
+			Move-Item -Path ($file + ".log") -Destination ($file + ".error.log") -Force
+			Write-Host ((Get-Date).ToString() + " Error encountered while running " + $file + ".sas")
+			exit $LastExitCode
+		}
+		return (ls ($file + ".log")).LastWriteTime
+	}
+	else {
+		return Get-Date
 	}
 }
 
@@ -19,20 +33,22 @@ foreach($file in Get-Content $build) {
 	if (!(Test-Path ($file + ".sas"))) {
 		throw [System.IO.FileNotFoundException] ($file + ".sas not found")
 	}
+	elseif ($force) {
+		$lastlogdate = runsas -file $file
+	}
 	elseif (!(Test-Path ($file + ".log"))) {
-		runsas -file $file
+		$lastlogdate = runsas -file $file
 	}	
 	elseif ( (ls ($file + ".sas")).LastWriteTime -gt (ls ($file + ".log")).LastWriteTime ) {
-		runsas -file $file
+		$lastlogdate = runsas -file $file
 	}
 	elseif ($firstfile -eq 0) {
 		if ($lastlogdate -gt (ls ($file + ".log")).LastWriteTime) {
-			runsas -file $file
+			$lastlogdate = runsas -file $file
 		}
 	}
 	
 	$firstfile = 0
-	$lastlogdate = (ls ($file + ".log")).LastWriteTime
 
 }
 Write-Host ((Get-Date).ToString() + " Build Ended")
